Title: 두 수의 합, Time: 88ms, Memory: 4MB, Status: Runtime Error - CodeTree

+++ 240902/두 수의 합/sum-of-two-num.cpp
@@ -3,7 +3,7 @@
 #include<limits.h>
 using namespace std;
 #define ll  long long
-map<int, short> m;
+map<ll, int> m;
 
 ll factorial(int n) {
     ll ans = 1;
